Comment View, edit, create, delete, Added edited bool to model
Comment functionality in a separate view with a comment counter in original view (This will probably change to include max of three comments on the feed with option to view more. Edited bool added to post model and comment model to show which comments and posts have been edited in the view with an edited tag.